Create logic for CreatePrizeForm
Create the logic for creating a new prize for a given tournament. The
form validate write a name for the prizem the format for the prize
amount and prize percent.

Implement and interface IPrizeRequester in the CreateTournament form so
the tournament form can get the prize created for the CreatePrize form.