Finish form to create new team
The combo box to get the people from the database is filled and when the
button to create a team is clicked the team is inserted in database with
the team members.

There is a validations for put a name to the team, add team members of
to the team and in case a new team member is in process to be created.